requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
brotli==1.2.0
certifi==2026.1.4
click==8.3.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.133.0
ffmpy==1.0.0
filelock==3.24.3
fsspec==2026.2.0
gradio==6.6.0
gradio-client==2.1.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.13.0
jupyter-client==8.8.0
jupyter-core==5.9.1
llama-cpp-python==0.3.16
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.23.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt-toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
rich==14.3.3
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.52.1
tomlkit==0.13.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typer-slim==0.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uvicorn==0.41.0
wcwidth==0.6.0
